pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["Dmitry Zemlaynkin <ygyshka0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.9"
coverage = "^7.3.4"
flake8 = "^6.1.0"
python-dotenv = "^1.0.0"
django-filter = "^23.5"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
